Configuration of DMS schema components CDF upload (#247)
[0.60.0] - 30-01-24
Added
Configuration for which DMS schema components are to be uploaded to CDF
Configuration for which DMS schema components are to be removed to CDF
Configuration how to handle existing CDF schema components during upload
Changed
Renamed UploadDMSDataModel to ExportDMSSchemaComponentsToCDF step. Note this is a breaking change, but
since we are on 0. version, we can do this.
Renamed DeleteDMSDataModel to DeleteDMSSchemaComponents step. Note this is a breaking change, but
since we are on 0. version, we can do this.
Renamed ExportDMSDataModel to ExportDMSSchemaComponentsToYAML step. Note this is a breaking change, but
since we are on 0. version, we can do this.
Renamed DataModel class to DMSSchemaComponents to better reflect the content of the class. Note this is a breaking change, but
since we are on 0. version, we can do this.
Step that waits for human approval timeout set to 1 day