feat: Add IMF rules importer (#552) (#571)
* feat: Add IMF rules importer (#552)

* feat: Add IMF rules importer

A ned step for importing rules has been added. The step converts
Information Modelling Framework (IMF) types, given as SHACL shapes,
into rules.

The new importer follows the compliance rules set by the OWL Importer,
but most of the logic are, in this case, included the SPARQL queries.
However, the validation checks are kept to ensure conformance.

The IMF identifiers are given as UUIDs, but CDF does not allow identifiers
starting with a number. All the relevant identifiers are therefore prefixed
with "IMF_"

IMF - attributes are properties that are described with several triples.
An attribute contains relations to reference data describing the meaning
of the attribute, as well as potential qualifiers and units of measure.
To ensure that this information is included the attributes are created as
both a class and property. This punning is required to ensure that both
the usage of the attribute relative to an IMF block or terminal and the
relationship to relevant reference data is captured.

* fix: Update value type

In cases where a sh:property points to a sh:class or a sh:qualifiedValueShape
the corresponding value is set as the expected value type of the property.

The logic should also cover rdfs:range.

---------

Co-authored-by: Tvedt <bard.tvedt@akersolutions.com>

* update imports

* make joined rdf based importer module

* make joined rdf based importer module

* made shared methods

* final fix

* bump version

* fix import

---------

Co-authored-by: Bård Henning <bhtvedt@gmail.com>
Co-authored-by: Tvedt <bard.tvedt@akersolutions.com>