fix: update identities created by IMF-importer
Main fix:
The IMF identifiers are GUIDs. In the original feature these identifiers
were prefixed with IMF_ to meet the requirement that the identifiers
shouldn't begin with a number.

An additional requirement is that the identifiers don't contain hypens.
In this commit the hypens are replaced by underscore.

Minor updates:
- Attributes are types both DatatypeProperty and AttributeType. This caused
the previous version to create duplicate entries for the attributes,
one with propertyType DatatypeProperty and the other with ObjectProperty.

The SPARQL has been updated to remove the duplication.

- imf:Attributes are objects, i.e. they have their own triples carrying
information about their meaning. The main predicate linked to an
imf:AttributeType is the imf:predicate, which points to a base reference
term. The imf:predicate is required to create an attribute, and the
cardinality was therefore set to 1-1. However, when the spreadsheet was
reviewed the meaning became that a block required the attribute. The
latter interpretation is not correct and therefore the cardinality of
the imf:predicate has been returned to the default 0-1.