[NEAT-801] New release process (#1019)
# RISK REVIEWER FYI

The same process has already been approved for Toolkit:
https://github.com/cognitedata/toolkit/pull/1487

# Description

After a discussion with the risk-reviewer team during the office hours
26/02, we are recommended to change our release process.

The changes are as follows:
* Changelog is no longer in the code base - only available in GitHub
releases. This is to avoid unnecessary merge conflict of the changelog.
* A changelog entry should be part of the PR description. The PR
description is now expected to follow the template shown below.
* The release process controls the versions, so the placeholder `0.0.0`
is used in the main branch and only replaced just before release.


Note this depends on 

![image](https://github.com/user-attachments/assets/c2332f35-c344-4d9c-b9f9-cf923210fe8c)


## Bump

- [x] Patch
- [ ] Minor
- [ ] Skip

## Changelog
### Added
- This will be ignored as I have checked skipped above.