🏋️‍♂️DMSExtractor use view external ID as rdf.type (#1039)
# Description

In the uplift workflow, we expect he type that is read from a view to
match the view externalId. If `node.type` is set, this assumptions can
be broken. This PR uses the view external ID as the rdf type.

## Bump

- [x] Patch
- [ ] Minor
- [ ] Skip

## Changelog
### Fixed

- When reading instances using `neat.read.cdf.graph(...)`, neat now
enforces the instances type to match the view external ID.