[NEAT-839]🤺OpenPyXL warning showing up as NeatValueWarning (#1051)
# Description

The following warnings show up when you read a spreadsheet that is
produced with Neat. The reason is the new dropdown feature, what is not
supported to be read back by `Openpyxl`. This is not an issue for Neat
as it validates the dropdowns up on read.

## Bump

- [x] Patch
- [ ] Minor
- [ ] Skip

## Changelog
### Fixed

- Reading a DMS spreadsheet produced by neat no longer produces a
`NeatValueWarning: UserWarning: Data Validation extension is not
supported and will be remove` for each property in the sheet.