[NEAT-837] DataSetId as URIRef and not Literal (#1048)
# Description
**REVIEWER** This is a ~10 line fix, with ~20 line test, and a 1800 line
snapshot YAML generated from the test.

## Context
When reading an asset centric graph `neat.read.classic.graph(...)`, we
fetch all assets, events, timeseries, sequences, and file metadata
connected to an asset hierarchy. This also includes assets in a
different hierarchy. When we include assets in a different hierarchy, we
need to create the instance ID of the parent of that asset. To do this,
the `neat.read.classic.graph(...)` looks up the parent and stores the
`dataSetId` which is used to set the space. This was unfortunately
stored as a `Literal` (rdf terminology) and not a `URIRef`. This is
fixed with this PR.

## Before

![image](https://github.com/user-attachments/assets/bb0197e0-c1a8-4828-95aa-8413fad2e173)

## After

![image](https://github.com/user-attachments/assets/a9a31dd9-b0f7-4b6f-a987-f6af5ed0ba10)



## Bump

- [x] Patch
- [ ] Minor
- [ ] Skip

## Changelog
### Fixed

- The `neat.read.classic.graph("my_root_asset", identifier="id")` no
longer raises a `TypeError` when a relationship is connected to an asset
outside the hierarchy.