fix: Capture correct information for IMF-types.
An IMF_base_model has been introduced in CDF to capture core information
about IMF-types as well as provide the link to the IMF_Standard_model.

This commit updates the classes and properties extracted from the IMF-
types, so that it is in accordance with the IMF_base_model. It also
include a mapping from:
    imf:Terminal -> imf:base:Terminal(version=v1)
    imf:Block -> imf:base:Block(version=v1)

Minor additional changes are:
- Custom strategy for making identifiers CDF compliant
- IMF related metadata