NEAT-858 Missing labels in ontology (#1066)
# Description

When creating ontology out of DMS rules we were not properly creating
triple that holds display `name` of `View` and `property`. The issue was
on one part in conversion from DMS rules to Information rules, and then
proper handling of creation of triple that hold `rdfs:label` and
`dcterm:title`.

## Bump

- [X] Patch
- [ ] Minor
- [ ] Skip

## Changelog
### Fixed
- Conversion of `name` from DMS rules to Information rules
- Conversion of `name` from Information rules to Ontology triple for
classes and properties

### Added
- Creation of triple holding information on title which generated by
combining external id and name