[NEAT-886] 👮 NeatGraphStore.transform catch issues (#1083)
# Description

This is part of a refactoring to NeatState, #1077. This modifies the
`.transform()` method of `NeatGraphStore` to return an `IssueList`. The
motivation for this is that this should comply with the other action
methods `NeatRuleStore.import_rules/transform()` and
`NeatGraphStore.write` that already return an `IssueList`. In addition,
it now returns an error if the graph transformation failed. This is
important as a failed transformation execution should not lead to a
state change.

## Bump

- [ ] Patch
- [ ] Minor
- [x] Skip