[prep ] Refactor inference (#1160)
# Description

Currently in the `neat.infer()` there is a standardization of property
names that is done to merge properties that have a slightly different
spelling.

This causes problem were I want to use the inference without doing this
standardization (working with Japanese characters that cannot be
standardized to `[a-zA-Z0-9]`.

I fix this by moving out the standardization logic to a separate
transformer that I call in the `neat.infer()` method.

## Bump

- [ ] Patch
- [ ] Minor
- [x] Skip

fix: imports