[prep NEAT-920] 📊 GraphImporter (#1183)
# Description

Rewrite of the `SubclassInferenceImporter` to `GraphImporter`.

This is to avoid merge conflicts with the main code base as this will be
a larger refactoring. The motivation for the refactoring is that the
current `SubclassInferenceImporter` is doing too much. It is both
inferring and updating an existing model. This will focus only on
inferring schema and move the logic for updating/changing an existing
model out to a transformer.

Difference from `SubclassInferenceImporter`

* Does not support existing data model. This is assumed to be handled by
a merge transformer.
* Does not find shared properties for subclasses. This is assumed to be
handled by a transformer.
* Does not use default prefixes. It only returns the prefixes actually
used in the data. Default prefixes can be added later.

## Bump

- [ ] Patch
- [ ] Minor
- [x] Skip