Refactor YAMLImporter to Dict Importer (#1168)
# Description

Refactoring YAMLImporter into DictImporter.This refactor enables to have
other "dict like" importers such as plain JSON files, and also Python
dict (not reading from a static file).

Also, It introduces a base clase for file read in the importers, which
can also be reused later on for spreadsheets or other sources.

I am skipping a version bump for now as there is no change to the end
users, but the upcoming PRs will introduce new functionalites which will
be possible because of this refactor

## Bump

- [ ] Patch
- [ ] Minor
- [x] Skip

## Changelog
### Improved
- Refactoring YAMLImporter into a generic DictImporter
- Having a BaseFileReader class so we can reuse this later