[NEAT-957] 🏴‍☠️ Statistics endpoint (#1163)
# Description

Context: We recently experienced an incident where Neat exceeded the
instance limit for a project, causing InField in production to stop
accepting new writes. This is the first part of the mitigation process,
which involves implementing support for checking the instance limit
using the NeatClient.

The reason this functionality is not included in the cognite-sdk is that
the endpoint is not yet ready and should not be exposed to end users
through the SDK. However, in this pull request, it is not directly
exposed to the end user. In the next pull request, I will implement a
check to be performed before loading instances into CDF.

## Bump

- [ ] Patch
- [ ] Minor
- [x] Skip