Generic context on Imported Data Model (#1258)
# Description

`ImportedDataModel` is an internal data structure we use to import data
models from different sources such as spreadsheet, CDF, ontologies and
so on. We use the context today when the source is a spreadsheet to tell
the user exactly which row have warnings/errors to make it easy for them
to fix.

This PR changes the type of context from one that is only valid for a
spreadsheet source to a more generic one. This is because in an upcoming
PR we will use the context in a different importer.

## Bump

- [ ] Patch
- [ ] Minor
- [x] Skip