[prep THIS-15] 😁 Return on failure (#1270)
# Description

**Context**: Neat has a concept call entities, a concept borrowed from
semantic data modeling - think of it is a noun. We use it to represent
view, containers, indexes and so on. When we import a data model from
any source, for example a spreadsheet, we need to defer error handling
of deserialization of entities such that we can do it in a proper
context.

This PR introduces a flag in the deserialization method of entities
(`.load()`) `return_on_failure` that can be set to defer error handling,
which we will use in a subsequent PR to defer error handling. We defer
error handling such that we can raise multiple errors simultaneously
with the right context (for example which row in a spreadsheet) to give
good error messages to the end-user of neat.

## Bump

- [ ] Patch
- [ ] Minor
- [x] Skip