Enabling restrictions in conceptual data model: adding restriction entities  (#1268)
# Description

This is first of a few PRs that are tackling importing of data model
restrictions often found in ontologies. PR brings definition of two
basic forms of restrictions:
- Concept property constraint on values
- Concept property constraint on cardinality and value type

In addition to the definition of the two new entities, a short form of
how those entities are represented as a string is provided.

## Bump

- [ ] Patch
- [ ] Minor
- [x] Skip

## Changelog
### Added

- Added new neat entities: `ConceptPropertyValueConstraint` and
`ConceptPropertyCardinalityConstraint` setting foundation for parsing of
OWL restrictions