Refactoring dm adding key entities datatypes (#1281)
# Description

This is **2** of a number of PR that are focused on refactoring neat
data model features to be exposed as Toolkit plugin. In this specific PR
important parts of _constants.py has been copied, _identifies.py has
been moved to better place, necessary pydantic models have been moved
from `_single_values.py` and `data_types.py` are moved and simplified to
_base.py and _data_types.py. Major change is dropping of parsing of
strings, as now we have dedicated AST for that (in review).

To Risk Reviewers: No new features added, only pieces of code moved to
the new module. Added roughly 150 lines of code for tests. This is
stacked PR!

## Bump

- [x] Patch
- [ ] Minor
- [ ] Skip

## Changelog
### Added

- Entities, data types, constants to new data_model module

---------

Co-authored-by: Anders Albert <60234212+doctrino@users.noreply.github.com>