[THIS-730] 🐋 Add container pydantic (#1284)
# Description

<del>Stacked on #1283, #1285</del>

**Context**: We are changing the internal structure of the neat data
models. We are dropping the physical data model format in Neat, and
instead will have a one-to-one DMS format identical to the API. In
principle, we could use the data classes that are in the `cognite-sdk`,
but we decided to reimplement them in `pydantic`. This is to 1. Get
better validation. 2. Better performance on serialization (pydantic does
serialization in rust), 3. Make them easy to extend and connect to the
Neat Conceptual model.

This PR implements the Pydantic class for Container. 

Note this PR has very few tests. This is because this an implementation
of the API in pydantic. Having a lot of tests of the
serialization/deserialization is essentially testing that `pydantic`
functionality. What makes sense to test is the error messages, and
ensure that they are adapted to human readable messages as the Neat user
base is data modelers and not necessarily developers. I have moved that
logic and testing out to #1286.

## Bump

- [ ] Patch
- [ ] Minor
- [x] Skip