[THIS-756] 👮 View tests (#1308)
# Description

**Context**: We are changing the internal structure of the neat data
models. We are dropping the physical data model format in Neat, and
instead will have a one-to-one DMS format identical to the API. In
principle, we could use the data classes that are in the `cognite-sdk`,
but we decided to reimplement them in `pydantic`. This is to 1. Get
better validation. 2. Better performance on serialization (pydantic does
serialization in rust), 3. Make them easy to extend and connect to the
Neat Conceptual model.


This PR implements validation for views. 

I also introduce a deviation from the view API. I set `connectionType`
on primary properties, this is to use discriminator on `connectionType`
to select the correct property type.

## Bump

- [ ] Patch
- [ ] Minor
- [x] Skip