THIS 752: Correct target class and value type setting (#1304)
# Description

This is stacked PR on top of #1303 

During the testing of SHACL engine, which was taking shacl rules
generated by neat, it has beed detected that we were setting
`NodeShape`'s `targetClass` and `PropertyShape`'s `node` to URI of SHACL
shapes. The correct way is to set those properties to class URI, which
in our case is stored as `instance_source` (if conceptual data model is
linked to the instances) or to generate such URI from concept entity
suffix and data model namespace. Furthermore, `PropertyShape`'s `path`
was not leveraging `instance_source` of property (if it exists). The
above has been implemented in this PR.


## Bump

- [x] Patch
- [ ] Minor
- [ ] Skip

## Changelog
### Improved

- Node shape target class and property shape node are now set to
instance source (if exists) or class uri
- Property shape path is now set to instance source (if exists) or class
uri when property is object property