THIS-268 Add Concept Pydantic Model (#1306)
# Description

This is stacked PR on top of #1301

`Concept` pydantic model has been added. The model shape is similar to
DMS `View` pydantic representation. This has been purposely done as it
will aid simpler conversion between DMS and Conceptual representation

No tests as we are trusting Pydantic doing what is suppose to do

## Bump

- [x] Patch
- [ ] Minor
- [ ] Skip

## Changelog
### Added

- Add Concept Pydantic Model