[THIS-773] 👷‍♂️ Added fixes to DMS pydantic (#1325)
# Description

Context We are in the process of reimplementing the core part of Neat.
The previous PRs have introduced a new internal structure (1-to-1 with
API) for DMS classes.

I have started to use the newly created pydantic classes for the DMS API
in #1323. I found several issues that are fixed with this PR.


**Note**: This code has not yet been released thus no release is
necessary.

## Bump

- [ ] Patch
- [ ] Minor
- [x] Skip