[THIS-777] 📂 Add source (#1332)
# Description

Stacked on <del>#1327 and #1328</del>

**Context** We are in the process of reimplementing the core part of
Neat.

This PR is a a split from #1323. It adds a data class for storing
information about the source of the the table was read from. This is
used in the error message of the upcoming importer such that the user
can get information about exactly which sheet, row, column that caused
the issue.

## Bump

- [ ] Patch
- [ ] Minor
- [x] Skip