[THIS-747] 👷‍♂️Table importer (#1334)
# Description

**Context** We are in the process of reimplementing the core part of
Neat.

This PR is a a split from #1323. It implements the logic for reading a
DMS data model from a Table format into the DMS API format. Note this is
a big PR as we are converting the entire Table format to DMS format. We
want test cases that are complete Table format of DMS converted to
schemas, thus will be large.

## Bump

- [ ] Patch
- [ ] Minor
- [x] Skip