[THIS-840]  🧱 Item Differ (#1363)
# Description

**Context** This is part of the rewrite of Neat for v1, and this one PR
in the series to add support for writing data models from Neat to CDF.

This PR introduces `ItemDiffer` and implements it for `SpaceRequest`.
This implementation will be used in the deployment to compare cdf vs
neat data model and determined whether a deployment is safe, and exactly
what has changed.

## Bump

- [ ] Patch
- [ ] Minor
- [x] Skip