[THIS-839] 🔧 Simplify http client (#1365)
# Description

**Context** This is part of the rewrite of Neat for v1, and this one PR
in the series to add support for writing data models from Neat to CDF.

This PR refactors the `HTTPClient` of Neat to prepare for the Schema
Deployer. I have done the following changes:

1. `ResponseMessages` now stores a body of string instead of bytes,
[ref](https://github.com/cognitedata/neat/pull/1365/files#diff-60e8b061c813dcf0914ca6c0574a847e8ac32813d8777aff6dc869a35591081bR94)
This is because the `httpx.Response` text property is decoded with the
encoding specified in the response. If we use bytes we lose information
about which encoding was used.
2. `ItemMessages` are now for multiple items instead of a single,
[ref](https://github.com/cognitedata/neat/pull/1365/files#diff-60e8b061c813dcf0914ca6c0574a847e8ac32813d8777aff6dc869a35591081bL120).
This enable use to track exactly which request items failed the request,
failed response and were successful:
[ref](https://github.com/cognitedata/neat/pull/1365/files#diff-60e8b061c813dcf0914ca6c0574a847e8ac32813d8777aff6dc869a35591081bR132).
I need this in the `SchemaDeployer` to know the result of each request
for each item.
3. Simplified the `ItemsRequest` such that it no longer takes in a
callable,
[ref](https://github.com/cognitedata/neat/pull/1365/files#diff-60e8b061c813dcf0914ca6c0574a847e8ac32813d8777aff6dc869a35591081bL162).
Instead, I moved out the `ReferenceObject` from the DMS model and assume
that items in a request can be turned onto a reference object,
[ref](https://github.com/cognitedata/neat/pull/1365/files).
4. Introduced `SpaceReference`
[ref](https://github.com/cognitedata/neat/pull/1365/files). This is such
that I can standardize the `as_reference` methods in the pydantic object
for DMS. This simplifies the neat.space.retrieve` method,
[ref](https://github.com/cognitedata/neat/pull/1365/files#diff-920fe2e55c56b4caf41bdd0cc8d21a292a8c79ca873ad633a1f456839b1d8a71R31)

## Bump

- [ ] Patch
- [ ] Minor
- [x] Skip