[prep THIS-856] 👨‍💻 Full field paths (#1386)
# Description

This PR refactors the Differs to return full field path for each field.
For example, if the the type has changed from `text` to `int32` for
property `myProperty` in a container, the field path will now be
`properties.myProperty.type.type` instead of only `type`. This will make
the diffs easier to display to the end user.

## Bump

- [ ] Patch
- [ ] Minor
- [x] Skip