[THIS-844] 🦂Parsing default prefix (#1440)
# Description

Handles if there is containers/views specified with/without default
space/version in Properties/Views/Container sheet.

Note - if a view is not found for a view property we raise an Error. If
not this row will be silently ignored. However, for containers, we allow
this as the user might reference a container in CDF, which checked by a
validator.

## Bump

- [x] Patch
- [ ] Minor
- [ ] Skip

## Changelog
### Fixed

- [alpha - v1] Neat can now read spreadsheets when default space is
specified for views/containers.