Property sparql query update (#1436)
# Description

1) Added additional optional part of query to get the domain when a
property is a subproperty of another property with domain given.

E.g if the .ttl files has these properties:

```
def-ops:fulfilledByDelivery a owl:ObjectProperty ;
    rdfs:range def-ops:MaterialLocationChangeProcess ;
    rdfs:subPropertyOf def-ops:fulfilledBy .

def-ops:fulfilledBy a owl:ObjectProperty ;
    rdfs:domain def-ops:Order .
```

In the current version fulfilledByDelivery will not have a domain and it
will show up in the excel sheet as "#N/A".
With the current version, it will get the same domain as the parent
property (here def-ops:Order)

2) Changed the maxCount and minCount to max_count and min_count so these
fields are correctly entered as "Max count" and "Min Count " in the
excel sheet. (Currently this is not the case, as it either looks for the
alias "Max Count" or "max_count" as given by the data class`
UnverifiedConceptualProperty`

## Bump

- [x] Patch
- [ ] Minor
- [ ] Skip

## Changelog
### Improved

- Updated property SPARQL query to get and use the domain of the parent
property if not given on the property itself.

---------

Co-authored-by: Nikola Vasiljevic <35523348+nikokaoja@users.noreply.github.com>