[THIS-933] 🤨Change what is breaking (#1476)
# Description

The API does not consider these changes breaking, however, I suggest we
override these as they will cause for example Pygen to break. Notice
also that if we want to reverse a change from Breaking-> Warning in the
future we can do that after v1, as the user will still be able to deploy
whatever models they have. However, redefining a change from Warning->
Breaking is a breaking change and would in principle require a major
bump to do. Thus if we decide to reverse this in the future, it is
possible after v1, but if we wait the reverse is not possible.

## Bump

- [x] Patch
- [ ] Skip

## Changelog
### Changed

- [alpha v1] Neat now consider changing an edge `direction` or
`edgeSource` a breaking change. In addition, if you change the
`collocation` of a text property in a container it is also breaking.