EDGE CASE: Incomplete Extension of Core Data Model (#1522)
# Description

Issue we had previously is that we were considering only views defined
in either resources.merged.views when we are fetching
resources.properties_by_view. However, this does not cover edge case
where view properties contain reverse direct on a view that is not part
of the current data model.

This is often the case of when core data model is partially extended,
i.e. when it does not include all the views.
Consequence is that we are raising Consistency Error
`NEAT-DMS-CONNECTIONS-REVERSE-002` which should not be raised as seen in
the following example:

<img width="1787" height="742" alt="image"
src="https://github.com/user-attachments/assets/06995492-0dd2-47ea-bcd4-60c8c9276042"
/>


Accepting that [the straw-man 2 and straw-man
3](https://people.well.com/user/doctorow/metacrap.htm) are always true,
thus we do not trust users doing data modeling to do it properly, now
expansion of properties has been improved.

Due to the same, finding right ancestory tree is consequently improved
as well.

## Bump

- [x] Patch
- [ ] Skip

## Changelog
### Improved

- View property expansion and ancestor tree generation

---------

Co-authored-by: Anders Albert <60234212+doctrino@users.noreply.github.com>